requirements.txt
pandarallel
zemberek-python
pandas